\name{performanceEval.GenoudBP}
\alias{performanceEval.GenoudBP}
\title{
performance evaluation for GenoudBP}
\description{
an implementaion for preformanceSEval with GenoudBP object}
\usage{
performanceEval.GenoudBP(model, x, y, scaled = F, dataLabel)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{
a genoudBP object
}
  \item{x}{
input data
}
  \item{y}{
input data}

}


\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
\code{\link{rocDisplay}}
\code{\link{predict.GenoudBP}}
}
