\name{performanceEval}
\alias{performanceEval}
\title{
performance evaluation}
\description{
this is the general function for performance evaluation}
\usage{
performanceEval(object, ...)
}

\arguments{
  \item{object}{
 the model to be evaluated
}
  \item{\dots}{
 other parameters for different implemetation

}
}
